requirements.txt
discord.py==2.3.2
python-dotenv>=0.21.0
pydantic>=1.10.0
aiohttp>=3.8.4
PyNaCl>=1.5.0
yt-dlp>=2023.12.0
